Updated readme.md example
Hey, just wanted to thank you for posting this driver! I was playing around with some things over the holidays and was happy to see someone ported this to Go! Just wanted to simplify the readme example a little. You don't have to include this, but I was wondering what the Interval was at first glance, just wanted to make it obvious. Cheers. Also, I'm porting the amg8833 driver right now based off your driver. So, a double thank you.